+++ src/unix/notbsd/linux/musl/b32/asmjs.rs
@@ -1,0 +1,2 @@
+pub type c_char = u8;
+pub type wchar_t = u32;
