+++ src/unix/notbsd/linux/musl/b32/x86.rs
@@ -1,0 +1,2 @@
+pub type c_char = i8;
+pub type wchar_t = i32;
